ArC: fix treatment of the "current" injection point
This mainly fixes a bug with `BeanManager.getReference()`, but it also
attempts to cleanup the `InstanceImpl` implementation and adds some
more tests.

This commit would also remove a TCK exclusion for `InjectionPointTest`,
except the test uses stereotype annotations on injection points, which
we consider an error. There's a challenge for it.