ArC: fix too strict circular dependency detection
ArC used to treat producers as depending on their declaring beans, so if a bean
also has a dependency on one of the producers it declares, circular dependency
exists. This is too strict for `static` producers: they in fact don't depend
on their declaring beans, so no circularity exists in such case. Circular
dependency only exists if the producer is not `static`. This commit fixes that.

When creating the `Bean` instance of a bean that depends on another bean, ArC
either supplies the `Bean` instance of the dependency directly, in case it already
exists, or it allows deferred instantiation in most cases. There used to be one
case that didn't allow deferred instantiation of `Bean` objects: a declaring bean
of a producer. There is no reason for that, deferred instantiation works correctly
even in this case, so this commit allows it.