ArC: prevent recursive interception in the strict mode
This commits adds a `CurrentInterception` class that maintains a thread-local
stack of methods that are currently being intercepted. If a method is about
to be intercepted, but it is already present on this stack, interception
is skipped and the method is invoked directly. Since recursive interception
actually happens extremely rarely in practice and the detection is costly,
this only happens in the strict mode. By default, all this is skipped and
the `CurrentInterception` class shouldn't even be loaded to the JVM.