ArC: fix inheriting method-level interceptor binding annotations
Method-level annotations are never inherited. This means that the lowest
override of a method determines whether method-level interceptor binding
annotations are present.

ArC used to behave differently. If a method on a class had no interceptor
binding annotations but overrode a method from a superclass that _did_
have interceptor binding annotations, the method-level annotations from
the superclass were inherited. This commit fixes that.